Use grpc.WithBlock() to wait for grpc connections to come up. (#6)
* Use grpc.WithBlock() to wait for grpc connections to come up.

We were rolling our own way of doing this via WaitForStateChange on
the connection object. But we could leak that connection object if
we didn't go directly from "connecting" to "idle", as the code
expected. It's cleaner just to use the WithBlock() option and let
grpc determine when we're connected.

* Added TestClientConnect

This test starts a client and then three seconds later starts a
grpc server for the client to connect to. The expected behavior
is that the connect process blocks until the server is up.

* Added TestClientConnect

This test starts a client and then three seconds later starts a
grpc server for the client to connect to. The expected behavior
is that the connect process blocks until the server is up.

* Added comment in test about potential race