Support unindexed labels (#12)
* Support unindexed labels

I thought it would be good to keep the metricd client API as
backwards-compatible as possible. So I changed LabelSet to
be a map[string]interface{} instead of a map[string]string.
So old LabelSet literals will continue to be compatible with
the new API. The proto changes also make it so that old clients
can talk to new metricd servers although this shouldn't
matter as much.

The value of the LabelSet is either a string (old behavior:
indexed label) or an UnindexedLabel instance (new behavior:
unindexed label). All other types will throw an error.

The rest was mostly piping this through and adding tests.

* Updated naming

protobuf "labels" field --> "indexed_labels"
metricd.UnindexedLabel --> metricd.Unindexed